添加不规格label的demo

+++ iOS_Demo/Demos/比例拼图/Controller/CQJigsawViewController.m
@@ -50,7 +50,7 @@
 
 // cell点击
 - (void)categoryView:(CQCategoryView *)categoryView clickedCellAtIndex:(NSInteger)index {
-    NSLog(@"点击第%ld个cell", index);
+    NSLog(@"点击第%ld个cell", (long)index);
 }
 
 @end
